move contract converters to subscriptionsapi file
