+++ control-plane/pkg/contract/extensions_test.go
@@ -19,10 +19,7 @@ package contract_test
 import (
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/stretchr/testify/require"
-	v1 "knative.dev/eventing/pkg/apis/eventing/v1"
 
 	"knative.dev/eventing-kafka-broker/control-plane/pkg/contract"
 )
@@ -32,105 +29,4 @@ func TestContractIncrementGeneration(t *testing.T) {
 	c.IncrementGeneration()
 
 	require.Equal(t, uint64(43), c.Generation)
-}
-
-func TestFromSubscriptionFilter(t *testing.T) {
-	tt := []struct {
-		name      string
-		apiFilter v1.SubscriptionsAPIFilter
-		want      *contract.DialectedFilter
-	}{
-		{
-			name:      "empty all filter",
-			apiFilter: v1.SubscriptionsAPIFilter{},
-			want:      &contract.DialectedFilter{},
-		},
-		{
-			name: "all filter with exact, prefix & suffix",
-			apiFilter: v1.SubscriptionsAPIFilter{
-				All: []v1.SubscriptionsAPIFilter{
-					{
-						Exact: map[string]string{"type": "dev.knative.kafkasrc.kafkarecord"},
-					},
-					{
-						Prefix: map[string]string{"source": "dev.knative"},
-					},
-					{
-						Suffix: map[string]string{"subject": "source"},
-					},
-					{
-						Not: &v1.SubscriptionsAPIFilter{
-							Prefix: map[string]string{"source": "dev.knative"},
-						},
-					},
-					{
-						SQL: "subject = 'source'",
-					},
-				},
-			},
-			want: &contract.DialectedFilter{
-				Filter: &contract.DialectedFilter_All{
-					All: &contract.All{
-						Filters: []*contract.DialectedFilter{
-							{
-								Filter: &contract.DialectedFilter_Exact{
-									Exact: &contract.Exact{
-										Attributes: map[string]string{"type": "dev.knative.kafkasrc.kafkarecord"},
-									},
-								},
-							},
-							{
-								Filter: &contract.DialectedFilter_Prefix{
-									Prefix: &contract.Prefix{
-										Attribute: "source",
-										Prefix:    "dev.knative",
-									},
-								},
-							},
-							{
-								Filter: &contract.DialectedFilter_Suffix{
-									Suffix: &contract.Suffix{
-										Attribute: "subject",
-										Suffix:    "source",
-									},
-								},
-							},
-							{
-								Filter: &contract.DialectedFilter_Not{
-									Not: &contract.Not{
-										Filter: &contract.DialectedFilter{
-											Filter: &contract.DialectedFilter_Prefix{
-												Prefix: &contract.Prefix{
-													Attribute: "source",
-													Prefix:    "dev.knative",
-												},
-											},
-										},
-									},
-								},
-							},
-							{
-								Filter: &contract.DialectedFilter_Cesql{
-									Cesql: &contract.CESQL{
-										Expression: "subject = 'source'",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
-			got := contract.FromSubscriptionFilter(tc.apiFilter)
-			if cmp.Diff(got, tc.want, cmpopts.IgnoreUnexported(contract.DialectedFilter{},
-				contract.All{}, contract.Prefix{}, contract.Suffix{}, contract.Exact{}, contract.Not{}, contract.CESQL{})) != "" {
-				t.Errorf("FromSubscriptionFilter() = %v, want %v", got, tc.want)
-			}
-		})
-	}
-
 }
